project/requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=12.0
requests>=2.31
PyYAML>=6.0
tabulate>=0.9
